Add Method for geting Leveled Items from a Unit
Method signature: public static GetLeveledItem(Unit, string)

This method is a generic method that will work for all leveled items
such as Diffusal Blade and Boots of Travel. It will also return the
item with biggest level if multiple items exist. This is usually not
needed but it is good to cover some edge cases.

GetDagon now calls the new method.